server/requirements.txt
fastapi
uvicorn[standard]
azure-cosmos
azure-identity
azure-search-documents
azure-ai-projects
python-dotenv
pydantic
httpx
